+++ examples/stockquotes/phase3/aggregator.py
@@ -1,9 +1,6 @@
 from __future__ import print_function
 import Pyro4
 
-Pyro4.config.SERIALIZER = 'pickle'
-Pyro4.config.SERIALIZERS_ACCEPTED.add('pickle')
-
 
 class Aggregator(object):
     def __init__(self):