+++ examples/stockquotes/phase3/viewer.py
@@ -4,9 +4,6 @@ import sys
 import Pyro4
 
 
-Pyro4.config.SERIALIZER = 'pickle'
-Pyro4.config.SERIALIZERS_ACCEPTED.add('pickle')
-
 if sys.version_info < (3, 0):
     input = raw_input