+++ src/Pyro4/util.py
@@ -556,8 +556,8 @@ try:
     else:
         ver = serpent.__version__
     ver = tuple(map(int, ver.split(".")))
-    if ver < (1, 7):
-        raise RuntimeError("requires serpent 1.7 or better")
+    if ver < (1, 8):
+        raise RuntimeError("requires serpent 1.8 or better")
     _ser = SerpentSerializer()
     _serializers["serpent"] = _ser
     _serializers_by_id[_ser.serializer_id] = _ser