+++ docs/source/changelog.rst
@@ -9,6 +9,7 @@ Change Log
   You had to turn the metadata feature off to be able to use the json serializer at all)
 - flame explicitly checks for pickle to be enabled instead of causing connection level errors
 - PYRONAME uri resolving now also uses the _pyroHmacKey set on the proxy (if any)
+- stockquotes tutorial doesn't actually require pickle anymore, so removed that from code and docs
 
 
 **Pyro 4.31**